requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
boto3=1.34.93
botocore=1.34.93
ca-certificates=2024.2.2=h56e8100_0
certifi=2024.2.2
cffi=1.16.0
charset-normalizer=2.0.12
click=8.0.3
cloudpickle=3.0.0
colorama=0.4.6
contourpy=1.1.1
cycler=0.12.1
dask=2021.11.2
deepliif=1.1.10
dill=0.3.8
dominate=2.6.0
filelock=3.13.1
fonttools=4.51.0
fsspec=2024.3.1
future=1.0.0
histoqc=0.0.post376+g5b79989
idna=3.7
imageio=2.34.1
importlib-resources=6.4.0
intel-openmp=2021.4.0
javabridge=1.0.19
jinja2=3.1.3
jmespath=1.0.1
joblib=1.4.0
jsonpatch=1.32
jsonpointer=2.1
kiwisolver=1.4.5
libffi=3.4.4=hd77b12b_0
llvmlite=0.36.0
locket=1.0.0
markupsafe=2.1.5
matplotlib=3.7.5
mkl=2021.4.0
mpmath=1.3.0
networkx=3.1
numba=0.53.1
numpy=1.22.4
opencv-python=4.5.3.56
openjdk=11.0.9.1=h57928b3_3
openslide-python=1.3.1
openssl=3.0.13=h2bbff1b_0
packaging=24.0
partd=1.4.1
pillow=10.3.0
pip=23.3.1=py38haa95532_0
pycparser=2.22
pyparsing=3.1.2
python=3.8.19=h1aa4202_0
python-bioformats=4.0.7
python-dateutil=2.9.0.post0
python-javabridge=4.0.3
pyvips=2.2.3
pywavelets=1.4.1
pyyaml=6.0.1
pyzmq=26.0.2
requests=2.26.0
s3transfer=0.10.1
scikit-image=0.18.3
scikit-learn=0.24.2
scipy=1.6.2
segment-anything=1.0
setuptools=68.2.2=py38haa95532_0
six=1.16.0
sqlite=3.41.2=h2bbff1b_0
sympy=1.12
tbb=2021.11.0
threadpoolctl=3.5.0
tifffile=2023.7.10
toolz=0.12.1
torch=2.3.0+cu121
torchaudio=2.3.0+cu121
torchfile=0.1.0
torchvision=0.18.0+cu121
tornado=6.4
typing-extensions=4.11.0
urllib3=1.26.18
vc=14.2=h21ff451_1
visdom=0.1.8.9
vs2015_runtime=14.27.29016=h5e58377_2
websocket-client=0.58.0
wheel=0.41.2=py38haa95532_0
zipp=3.18.1
